+++ alacritty_terminal/src/config/window.rs
@@ -1,5 +1,8 @@
-use crate::config::{failure_default, Delta};
+use crate::config::{
+    failure_default, from_string_or_deserialize, option_explicit_none, Delta, FromString,
+};
 use crate::index::{Column, Line};
+use crate::window::DEFAULT_NAME;
 
 #[serde(default)]
 #[derive(Deserialize, Debug, Clone, Default, PartialEq, Eq)]
@@ -33,8 +36,12 @@ pub struct WindowConfig {
     pub title: Option<String>,
 
     /// Window class
-    #[serde(deserialize_with = "failure_default")]
-    pub class: Option<String>,
+    #[serde(deserialize_with = "from_string_or_deserialize")]
+    pub class: Class,
+
+    /// GTK theme variant
+    #[serde(deserialize_with = "option_explicit_none")]
+    pub gtk_theme_variant: Option<String>,
 
     /// TODO: DEPRECATED
     #[serde(deserialize_with = "failure_default")]
@@ -115,5 +122,25 @@ impl Dimensions {
     #[inline]
     pub fn columns_u32(&self) -> u32 {
         self.columns.0 as u32
+    }
+}
+
+/// Window class hint
+#[serde(default)]
+#[derive(Deserialize, Debug, Clone, PartialEq, Eq)]
+pub struct Class {
+    pub instance: String,
+    pub general: String,
+}
+
+impl Default for Class {
+    fn default() -> Self {
+        Class { instance: DEFAULT_NAME.into(), general: DEFAULT_NAME.into() }
+    }
+}
+
+impl FromString for Class {
+    fn from(value: String) -> Self {
+        Class { instance: value, general: DEFAULT_NAME.into() }
     }
 }