+++ alacritty_terminal/src/window.rs
@@ -61,6 +61,9 @@ pub struct Window {
     windowed_context: glutin::WindowedContext<PossiblyCurrent>,
     mouse_visible: bool,
 
+    /// Keep track of the current mouse cursor to avoid unnecessarily changing it
+    current_mouse_cursor: MouseCursor,
+
     /// Whether or not the window is the focused window.
     pub is_focused: bool,
 }
@@ -151,9 +154,8 @@ impl Window {
         dimensions: Option<LogicalSize>,
     ) -> Result<Window> {
         let title = config.window.title.as_ref().map_or(DEFAULT_NAME, |t| t);
-        let class = config.window.class.as_ref().map_or(DEFAULT_NAME, |c| c);
 
-        let window_builder = Window::get_platform_window(title, class, &config.window);
+        let window_builder = Window::get_platform_window(title, &config.window);
         let windowed_context =
             create_gl_window(window_builder.clone(), &event_loop, false, dimensions)
                 .or_else(|_| create_gl_window(window_builder, &event_loop, true, dimensions))?;
@@ -165,8 +167,13 @@ impl Window {
         // Set OpenGL symbol loader. This call MUST be after window.make_current on windows.
         gl::load_with(|symbol| windowed_context.get_proc_address(symbol) as *const _);
 
-        let window =
-            Window { event_loop, windowed_context, mouse_visible: true, is_focused: false };
+        let window = Window {
+            event_loop,
+            current_mouse_cursor: MouseCursor::Default,
+            windowed_context,
+            mouse_visible: true,
+            is_focused: false,
+        };
 
         window.run_os_extensions();
 
@@ -240,8 +247,11 @@ impl Window {
     }
 
     #[inline]
-    pub fn set_mouse_cursor(&self, cursor: MouseCursor) {
-        self.window().set_cursor(cursor);
+    pub fn set_mouse_cursor(&mut self, cursor: MouseCursor) {
+        if cursor != self.current_mouse_cursor {
+            self.current_mouse_cursor = cursor;
+            self.window().set_cursor(cursor);
+        }
     }
 
     /// Set mouse cursor visible
@@ -253,11 +263,7 @@ impl Window {
     }
 
     #[cfg(not(any(target_os = "macos", windows)))]
-    pub fn get_platform_window(
-        title: &str,
-        class: &str,
-        window_config: &WindowConfig,
-    ) -> WindowBuilder {
+    pub fn get_platform_window(title: &str, window_config: &WindowConfig) -> WindowBuilder {
         use glutin::os::unix::WindowBuilderExt;
 
         let decorations = match window_config.decorations {
@@ -267,7 +273,9 @@ impl Window {
 
         let icon = Icon::from_bytes_with_format(WINDOW_ICON, ImageFormat::ICO);
 
-        WindowBuilder::new()
+        let class = &window_config.class;
+
+        let mut builder = WindowBuilder::new()
             .with_title(title)
             .with_visibility(false)
             .with_transparency(true)
@@ -275,17 +283,19 @@ impl Window {
             .with_maximized(window_config.startup_mode() == StartupMode::Maximized)
             .with_window_icon(icon.ok())
             // X11
-            .with_class(class.into(), DEFAULT_NAME.into())
+            .with_class(class.instance.clone(), class.general.clone())
             // Wayland
-            .with_app_id(class.into())
+            .with_app_id(class.instance.clone());
+
+        if let Some(ref val) = window_config.gtk_theme_variant {
+            builder = builder.with_gtk_theme_variant(val.clone())
+        }
+
+        builder
     }
 
     #[cfg(windows)]
-    pub fn get_platform_window(
-        title: &str,
-        _class: &str,
-        window_config: &WindowConfig,
-    ) -> WindowBuilder {
+    pub fn get_platform_window(title: &str, window_config: &WindowConfig) -> WindowBuilder {
         let decorations = match window_config.decorations {
             Decorations::None => false,
             _ => true,
@@ -303,11 +313,7 @@ impl Window {
     }
 
     #[cfg(target_os = "macos")]
-    pub fn get_platform_window(
-        title: &str,
-        _class: &str,
-        window_config: &WindowConfig,
-    ) -> WindowBuilder {
+    pub fn get_platform_window(title: &str, window_config: &WindowConfig) -> WindowBuilder {
         use glutin::os::macos::WindowBuilderExt;
 
         let window = WindowBuilder::new()