+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Block selection mode when Control is held while starting a selection
+- Allow setting general window class on X11 using CLI or config (`window.class.general`)
+- Config option `window.gtk_theme_variant` to set GTK theme variant
+- Completions for `--class` and `-t` (short title)
+- Change the mouse cursor when hovering over the message bar and its close button
 
 ### Fixed
 
@@ -20,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Cells sometimes not getting cleared correctly
 - X11 clipboard hanging when mime type is set
 - On macOS, Alacritty will now fallback to Menlo if a font specified in the config cannot be loaded
+- Debug ref tests are now written to disk regardless of shutdown method
 
 ## 0.3.3